requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-oauth-toolkit==1.7.1
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
